Main screens moved to routes folder.
